+++ scripts/repo/winesapos-build-repo.sh
@@ -40,6 +40,8 @@ makepkg_fn firefox-esr-bin
 makepkg_fn game-devices-udev
 makepkg_fn gamescope-session-git install
 makepkg_fn gamescope-session-steam-git
+makepkg_fn opengamepadui-bin install
+makepkg_fn opengamepadui-session-git
 makepkg_fn heroic-games-launcher-bin
 makepkg_fn hfsprogs
 makepkg_fn lightdm-settings